Transfer methods from Basic to Expr.
as_terms, as_ordered_terms, as_ordered only make sense for arithmetic
expressions, not for every possible kind of object.
